Return the original Promise instead of a proxied one
This entirely avoids the issue of either propagating the rejection or
catching them to prevent spurious errors.

If the user just dispatches the action and does nothing more with the
promise, errors will be caught and only the failure action will be
dispatched.

If the user attaches a .then handler to the promise after dispatching
it, then it is the user's responsibility to .catch the Promise
rejection. This is essentially forking the Promise chain, and rejections
propagate on each branch, so they have to be caught on each branch.

Promise resolution is specified to execute the .then handlers in order,
so by the time a user-attached .then runs, the action will have already
been dispatched.